temp remove halfing

+++ main.py
@@ -158,7 +158,9 @@ def main(args):
     logger = TensorBoardLogger(args.log_dir, name="ArTEMIS")
 
     # NOTE: bf16-mixed halves precision on operations when available; this is for speedup purposes
-    trainer = L.Trainer(precision="bf16-mixed", max_epochs=args.max_epoch, log_every_n_steps=args.log_iter, logger=logger)
+
+    # trainer = L.Trainer(precision="bf16-mixed", max_epochs=args.max_epoch, log_every_n_steps=args.log_iter, logger=logger)
+    trainer = L.Trainer(max_epochs=args.max_epoch, log_every_n_steps=args.log_iter, logger=logger)
     trainer.fit(model, train_loader)
 
     # Test the model with Lightning
